requirements.txt
aiohttp==3.9.1
python-dotenv==1.0.0
requests==2.31.0
dnspython==2.4.2
whois==0.9.7
ssl-checker==0.4.0
